app/requirements.txt
fastapi[all]
sqlalchemy
psycopg2
pyjwt
passlib[bcrypt]
pytest